+++ metadata.rb
@@ -13,7 +13,6 @@ depends "mysql", "~> 4.0"
 depends 'nginx', '< 2.4.4'
 depends "database", ">= 2.0"
 depends "build-essential", ">= 1.4"
-depends 'ssh_known_hosts', '>= 2.0'
 
 %w{ debian ubuntu centos redhat fedora scientific amazon windows }.each do |os|
   supports os