+++ spec/recipes/capistrano_spec.rb
@@ -1,8 +1,6 @@
 describe 'config-driven-helper::capistrano' do
   context 'with apache site configuration' do
     let(:chef_run) do
-      stub_data_bag('users').and_return([])
-
       ChefSpec::SoloRunner.new do |node|
         node.set['apache']['sites']['inviqa'] = {
           'capistrano' => {
@@ -62,8 +60,6 @@ describe 'config-driven-helper::capistrano' do
 
   context 'with nginx site configuration' do
     let(:chef_run) do
-      stub_data_bag('users').and_return([])
-
       ChefSpec::SoloRunner.new do |node|
         node.set['nginx']['sites']['inviqa'] = {
           'capistrano' => {
@@ -85,11 +81,9 @@ describe 'config-driven-helper::capistrano' do
       end
     end
   end
-  
+
   context 'with custom mode configuration' do
     let(:chef_run) do
-      stub_data_bag('users').and_return([])
-
       ChefSpec::SoloRunner.new do |node|
         node.set['nginx']['sites']['inviqa'] = {
           'capistrano' => {
@@ -140,35 +134,6 @@ describe 'config-driven-helper::capistrano' do
           mode: '0771',
         )
       end
-    end
-  end
-
-  context 'with users databag containing a deploy user' do
-    let(:chef_run) do
-      users = {}
-      users['deploy'] = {
-        'username' => 'deploy',
-        'groups' => ['deploy']
-      }
-
-      stub_data_bag('users').and_return(users.keys)
-      users.each { |item, user| allow(Chef::EncryptedDataBagItem).to receive(:load).with('users', item).and_return(user) }
-
-      ChefSpec::SoloRunner.new do |node|
-        node.set['capistrano']['known_hosts'] = ['github.com', 'example.org']
-      end.converge(described_recipe)
-    end
-
-    it "will manage deploy user accounts" do
-      expect(chef_run).to create_user('deploy').with(
-        gid: 'deploy'
-      )
-      expect(chef_run).to create_group('deploy')
-    end
-
-    it "will supply known_hosts files for user accounts" do
-      expect(chef_run).to append_to_ssh_known_hosts 'github.com'
-      expect(chef_run).to append_to_ssh_known_hosts 'example.org'
     end
   end
 end