+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 IMPROVEMENTS:
 
   * #56 Loosen cookbook constraints to support httpd 2.4
+
+CHANGES:
+
   * #57 Disable nginx https variable emulation by default
+  * #58 Fork off deploy user work from capistrano recipe to new cookbook
 
 ## 1.5.0 (10 November 2015)